+++ hydromt/model/components/tables.py
@@ -1,14 +1,13 @@
 """Tables component."""
 
-import glob
 import logging
-from os.path import basename, join
 from typing import TYPE_CHECKING, Dict, Optional, Union, cast
 
 import pandas as pd
 
 from hydromt.model.components.base import ModelComponent
 from hydromt.model.steps import hydromt_step
+from hydromt.readers import _expand_wildcard_path
 
 if TYPE_CHECKING:
     from hydromt.model.model import Model
@@ -27,7 +26,7 @@ class TablesComponent(ModelComponent):
     def __init__(
         self,
         model: "Model",
-        filename: str = "tables/{name}.csv",
+        filename: str = "tables/*.csv",
     ):
         """Initialize a TablesComponent.
 
@@ -38,7 +37,7 @@ class TablesComponent(ModelComponent):
         filename: str
             The default place that should be used for reading and writing unless the
             user overrides it. If a relative path is given it will be used as being
-            relative to the model root. By default `tables/{name}.csv` for this
+            relative to the model root. By default ``tables/*.csv`` for this
             component, and can be either relative or absolute.
         """
         self._data: Optional[Dict[str, Union[pd.DataFrame, pd.Series]]] = None
@@ -72,7 +71,7 @@ class TablesComponent(ModelComponent):
             )
             return
 
-        filename = filename or self._filename
+        filename = (filename or self._filename).replace("*", "{name}")
 
         kwargs.setdefault("index", False)
         kwargs.setdefault("header", True)
@@ -88,17 +87,28 @@ class TablesComponent(ModelComponent):
 
     @hydromt_step
     def read(self, filename: Optional[str] = None, **kwargs) -> None:
-        """Read tables at provided or default file path if none is provided."""
+        """Read tables at provided or default file path if none is provided.
+
+        Parameters
+        ----------
+        filename : str, optional
+            filename relative to model root. Should contain a * wildcard character
+            to read multiple files into the data dictionary. All files matching the
+            glob pattern defined by filename will be read. The filename without
+            extension will be used as the key in the data dictionary.
+            If None, the path that was provided at init will be used.
+        **kwargs:
+            Additional keyword arguments that are passed to the
+            `pandas.read_csv` function.
+        """
         self.root._assert_read_mode()
         self._initialize_tables(skip_read=True)
         logger.info("Reading model table files.")
         fn = filename or self._filename
-        filenames = glob.glob(join(self.root.path, fn.format(name="*")))
-        if len(filenames) > 0:
-            for fn in filenames:
-                name = basename(fn).split(".")[0]
-                tbl = pd.read_csv(fn, **kwargs)
-                self.set(tbl, name=name)
+        for path in _expand_wildcard_path(self.root.path / fn):
+            logger.info(f"Reading table from {path} into model as '{path.stem}'.")
+            tbl = pd.read_csv(path, **kwargs)
+            self.set(tbl, name=path.stem)
 
     def set(
         self,